etl/requirements.txt
python_dotenv==1.0.0
psycopg2-binary==2.9.9
python_dateutil==2.8.2
elasticsearch==8.10.0
backoff==2.2.1
redis==5.0.1
pydantic==2.4.2